urgent-drafting/pyproject.toml
[project]
name = "urgent-drafting-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for urgent news drafting capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "python-multipart>=0.0.6",
    "reuters-ai-assistant-mcp-hitl>=0.3.7",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.15.0",
]

[project.scripts]
urgent-drafting-mcp = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
